this is 1st commit.
note that the original settings of PoCo was too old (the specifying target os was 10.9).
so that, settings need to be up-to-date and sdk linked also is updated.
therefore, at this moment, PoCo contains some problems (but, I've been aware of almost all problems, and I'll fix these ones. so, don't worry).