Feature/ref#41 add feature copylayer (#50)
* feat: add button (icon) and menu; copy layer.

add button (icon) on PoCoLayerWindow and menu for feature of copy layer.

* feat: add feature; copy layer.

add a feature that is to copy layer.
the feature of copying layer has implemented since before. but, to invoke that feature, required the mouse and/or keyboard operation so far.
the purpose of this implementation is that copy current layer right above with menu or button.

* feat: implement the copy layer menu item.

close #41.