merge develop into main because release version 1.5.0. (#56)
* fix: upgrade settings and resources.

ref #21.

* fix: upgrade resources (2).

ref #21.

* fix: be support for arm-based mac.

ref #5.
to be support for mac of both intel-based and arm-based, revise xcodeproj.

* feat: add quick look extension. (#27)

add a feature that is to provide the quick look extension.

close #20.

* add schemes.

* fix: fix an issue (scrolling detail information view of palette). (#28)

* fix: fix an issue (scrolling detail information view of palette).

fix an issue that the detail information view of palette does not scroll correctly.
in addition, a few comments are tweaked.

* fix: unregister observer.

close #7.

* fix: remove extra margin. (#29)

to remove extra margin on TableView, override style when awake from nib.
I think that this is affected by default behaviour changed (Apps that link to previous macOS versions default to NSTableViewStylePlain).

close #8.

* tweak comment.

* fix: fix issues that titles of layer view are broken. (#30)

* fix: fix issues that titles of layer view are broken.

* fix typo.

close #9.

* fix: fix an issue that image is currpted. (#31)

ref #11.
fix an issue that after rotate and/or resize image, image is corrupt when edit using selection shape.
affected features; flip image (both horizontal and vertical), auto gradient, replace colour, and texture.

close #11.

* fix: fix an issue. (#32)

fix an issue that the blur feature (water drop) does not work correctly. and, even app crash may be happened.

- iterating pixelmap for mask was wrong (bad memory access).
- the drawing position did not locate within the drawing rectangle.

close #10.

* fix: fix an issue that some views can not redraw correctly. (#33)

fix an issue that some views can not redraw correctly.
in adiition, tweak coding styles.
NOTE THAT the sub-view canNOT redraw correctly when scrolling at present. so that issue on the sub-view windows is investigating now.

**therefore created pull request, but do not delete corresponding issue after merge.**

* fix: fix an issue that some views could not redraw correctly. (#34)

* fix: fix an issue that some views could not redraw correctly.

ref #6.
fix an issue that some views could not redraw correctly.
(changes in #33 are revoked. so that, reimplement the fixes.)

* fix typos.

close #6.

* Feature/ref#25 fix warnings of deprecated constants (#35)

* fix: fix warnings.

fix warnings due to deprecated constants; `NSOnState` and `NSOffState`.

* fix: fix warnings.

fix warnings due to the deprecated constants; NSCommandKeyMask, NSAlternateKeyMask, NSControlKeyMask, and NSShiftKeyMask.

* fix: fix warning.

to fix warning, change target version symbol.

close #25.

* chore: correct comment.

* fix: fix an issue. (#42)

* fix: fix an issue.

fix an issue that `setAcceptsTouchEvents` is deprecated.

* replace comment (translate).

close #38.

* fix: fix an issue (#43)

fix an issue of the deprecated method `validateMenuItem` of NSObject.
Apply `NSMenuItemValidation` protocol to `PoCoView`, `PoCoAppController`, and `PoCoLayerTableView`.
And, in `PoCoLayerTableView`, do not forward `validateMenuItem` to super class.

close #39.

* Feature/ref#36 and #37 fix deprecated archiver classes (#44)

* chore: tweak style and translate some comment into Japanese.

* feat: declaration of support for secure restorable state.

* feat: adopt NSSecureCoding protocol to fix deprecation.

because `NSArchiver` and `NSUnarchiver` are deprecated, `PoCoPenStyle` and `PoCoTilePattern` is warned.
To fix this issue, the archived data should adopt `NSSecureCoding` protocol.

* fix: fix an issue that `NSArchiver` and `NSUnarchiver` are deprecated.

close #36.
close #37.

* feat: implement the migration path against `NSUnarchiver` deprecation. (#46)

close #45.

* fix: not call legacy method; `[NSUserDefaults synchronise].

* fix: fix an issue. (#47)

* fix: fix an issue.

fix an issue that [NSApp `beginSheet:`] has marked the deprecated.

* tweak comment.


close #40.

* fix warnings.

* fix: fix an issue on auto gradient sheet. (#49)

fix an issue that the detail setting sheet on auto gradient panel can not return the cancel response.

close #48.

* Feature/ref#41 add feature copylayer (#50)

* feat: add button (icon) and menu; copy layer.

add button (icon) on PoCoLayerWindow and menu for feature of copy layer.

* feat: add feature; copy layer.

add a feature that is to copy layer.
the feature of copying layer has implemented since before. but, to invoke that feature, required the mouse and/or keyboard operation so far.
the purpose of this implementation is that copy current layer right above with menu or button.

* feat: implement the copy layer menu item.

close #41.

* Feature/ref#12 add feature revert mono pattern to default (#51)

* feat: add menu items; revert the pen styles and the tile patterns.

to revert a pen style and/or a tile pattern, menu items are added.

* feat: implement a feature to revert pattern.

implement a feature to revert pattern; pen style and tile pattern.

close #12.

* feat: implement code for linking UI with edit features.

* tweak project settings.

* tweak layout on perferences panel.

* fix: tweak each table colmn with on layer window. (#54)

tweak each table colmn with on layer window.
in addition, apply the dedicated custom class to each table column (in 'Base.lproj').

close #53

* chore: bump up version number; 1.4.0 to 1.5.0. (#55)

close #52.